working on modularization
