--- password_manager/auth.py
@@ -1,98 +1,0 @@
-import os
-import getpass
-import time
-import click
-from pathlib import Path
-from cryptography.fernet import Fernet
-from .constants import SESSION_FILE, SESSION_TIMEOUT, MASTER_PASSWORD_FILE
-from .utils import load_secure_key, set_permissions, decrypt_master_password
-
-def authenticate_user():
-    try:
-        username = os.getlogin()
-        if username == 'root':
-            username = getpass.getuser()
-    except Exception:
-        username = getpass.getuser()
-
-    choice = click.prompt('Choose authentication method: [P]assword/[F]ingerprint', type=str).lower()
-
-    if choice == 'p':
-        stored_master_password = decrypt_master_password()
-        password = click.prompt('Master Password', hide_input=True)
-        if password != stored_master_password:
-            click.echo("Authentication failed.")
-            exit(1)
-        else:
-            click.echo("Authentication succeeded.")
-            with open(SESSION_FILE, 'w') as f:
-                f.write(str(int(time.time())))
-    elif choice == 'f':
-        if os.uname().sysname == "Darwin":
-            if not authenticate_fingerprint_mac():
-                click.echo("Fingerprint authentication failed.")
-                exit(1)
-            else:
-                click.echo("Fingerprint authentication succeeded.")
-                with open(SESSION_FILE, 'w') as f:
-                    f.write(str(int(time.time())))
-        else:
-            click.echo("Fingerprint authentication is not supported on this system.")
-            exit(1)
-    else:
-        click.echo("Invalid choice.")
-        exit(1)
-
-def ensure_authenticated():
-    """Ensure the user is authenticated before proceeding."""
-    if not is_authenticated():
-        click.echo("You need to authenticate first.")
-        authenticate_user()
-    else:
-        with open(SESSION_FILE, 'w') as f:
-            f.write(str(int(time.time())))
-
-def is_authenticated():
-    if not SESSION_FILE.exists() or os.stat(SESSION_FILE).st_size == 0:
-        return False
-    with open(SESSION_FILE, 'r') as f:
-        timestamp = int(f.read().strip())
-    if time.time() - timestamp > SESSION_TIMEOUT:
-        os.remove(SESSION_FILE)
-        return False
-    return True
-
-def authenticate_fingerprint_mac():
-    """Authenticate the user using fingerprint on macOS."""
-    import objc
-    from Foundation import NSObject
-    from LocalAuthentication import LAContext, LAPolicyDeviceOwnerAuthenticationWithBiometrics
-    from ctypes import CDLL, c_void_p, c_long
-
-    context = LAContext.alloc().init()
-    success, error = context.canEvaluatePolicy_error_(LAPolicyDeviceOwnerAuthenticationWithBiometrics, None)
-    
-    if success:
-        click.echo("Please authenticate using your fingerprint...")
-        semaphore = CDLL('/usr/lib/system/libdispatch.dylib').dispatch_semaphore_create(0)
-        
-        def callback(_success, _error):
-            nonlocal authenticated
-            if _success:
-                authenticated = True
-            else:
-                click.echo(f"Fingerprint authentication error: {_error}")
-                authenticated = False
-            CDLL('/usr/lib/system/libdispatch.dylib').dispatch_semaphore_signal(semaphore)
-
-        authenticated = False
-        context.evaluatePolicy_localizedReason_reply_(
-            LAPolicyDeviceOwnerAuthenticationWithBiometrics,
-            "Authenticate to access password manager",
-            callback
-        )
-        CDLL('/usr/lib/system/libdispatch.dylib').dispatch_semaphore_wait(semaphore, c_long(-1))
-        return authenticated
-    else:
-        click.echo(f"Fingerprint authentication not available: {error}")
-        return False
