+++ password_manager/utils.py
@@ -1,30 +1,58 @@
 import os
-import base64
-import logging
-import threading
-from pathlib import Path
-from cryptography.fernet import Fernet
+import getpass
+from cryptography.fernet import Fernet, InvalidToken
 from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
 from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.backends import default_backend
-from .constants import SECURE_KEY_FILE, SALT_FILE, CURRENT_DIRECTORY_FILE, DATA_DIR, MASTER_PASSWORD_FILE
-from .password_operations import reencrypt_passwords
+from pathlib import Path
+from ctypes import CDLL, c_void_p, c_long
+import threading
+import logging
+import base64
 
 # Setup logging to redirect to a file
 logging.basicConfig(filename='vault.log', level=logging.DEBUG, format='%(asctime)s - %(levelname)s - %(message)s')
 logging.info("Logging initialized")  # Initial log message to verify logging setup
 
+# Setup directories and files
+DATA_DIR = Path.home() / ".password_manager" / "data"
+SECURE_KEY_FILE = Path.home() / ".password_manager" / "secure_key.key"
+MASTER_PASSWORD_FILE = Path.home() / ".password_manager" / "master_password.enc"
+SESSION_FILE = Path.home() / ".password_manager" / ".session"
+CURRENT_DIRECTORY_FILE = Path.home() / ".password_manager" / ".current_directory"
+SALT_FILE = Path.home() / ".password_manager" / "salt"
+DATA_DIR.mkdir(parents=True, exist_ok=True)
+SESSION_TIMEOUT = 3600  # Set session timeout to 3600 seconds (1 hour)
+
+if os.uname().sysname == "Darwin":
+    import objc
+    from Foundation import NSObject
+    from LocalAuthentication import LAContext, LAPolicyDeviceOwnerAuthenticationWithBiometrics
+
+libdispatch = CDLL('/usr/lib/system/libdispatch.dylib')
+libdispatch.dispatch_semaphore_create.argtypes = [c_long]
+libdispatch.dispatch_semaphore_create.restype = c_void_p
+libdispatch.dispatch_semaphore_wait.argtypes = [c_void_p, c_long]
+libdispatch.dispatch_semaphore_wait.restype = c_long
+libdispatch.dispatch_semaphore_signal.argtypes = [c_void_p]
+libdispatch.dispatch_semaphore_signal.restype = c_long
+
 # Global lock for synchronizing key operations
 key_lock = threading.Lock()
 
-def decrypt_master_password():
-    """Decrypt and return the master password using the secure key."""
-    secure_key = load_secure_key()
-    fernet = Fernet(secure_key)
-    with open(MASTER_PASSWORD_FILE, 'rb') as f:
-        encrypted_password = f.read()
-    password = fernet.decrypt(encrypted_password).decode()
-    return password
+def derive_key(password: str):
+    global fixed_salt  # Access the global variable
+    kdf = PBKDF2HMAC(
+        algorithm=hashes.SHA256(),
+        length=32,
+        salt=fixed_salt,
+        iterations=100000,
+        backend=default_backend()
+    )
+    derived_key = kdf.derive(password.encode())
+    derived_key_b64 = base64.urlsafe_b64encode(derived_key)
+    logging.debug(f"Derived key: {derived_key_b64}")
+    return derived_key_b64
 
 def load_secure_key():
     """Load the secure key from the file."""
@@ -61,77 +89,40 @@ def load_salt():
     with open(SALT_FILE, 'rb') as f:
         return f.read()
 
+
 # Load the salt (generate it if it doesn't exist)
 fixed_salt = load_salt()
 
-def derive_key(password: str):
-    global fixed_salt  # Access the global variable
-    kdf = PBKDF2HMAC(
-        algorithm=hashes.SHA256(),
-        length=32,
-        salt=fixed_salt,
-        iterations=100000,
-        backend=default_backend()
-    )
-    derived_key = kdf.derive(password.encode())
-    derived_key_b64 = base64.urlsafe_b64encode(derived_key)
-    logging.debug(f"Derived key: {derived_key_b64}")
-    return derived_key_b64
-
-def load_current_directory():
-    if CURRENT_DIRECTORY_FILE.exists():
-        with open(CURRENT_DIRECTORY_FILE, 'r') as f:
-            return Path(f.read().strip())
-    return DATA_DIR
-
-def save_current_directory(current_directory):
-    with open(CURRENT_DIRECTORY_FILE, 'w') as f:
-        f.write(str(current_directory))
-
-def get_password_file_path(domain, folder=None):
-    current_dir = load_current_directory()
-    if folder:
-        target_dir = (current_dir / folder).resolve()
-    else:
-        target_dir = current_dir
-    return target_dir / f"{domain}.pass"
-
-class CipherSingleton:
-    _instance = None
-    _lock = threading.Lock()
-
-    def __new__(cls):
-        if cls._instance is None:
-            with cls._lock:
-                if cls._instance is None:
-                    cls._instance = super(CipherSingleton, cls).__new__(cls)
-                    cls._instance._cipher_initialized = False
+def reencrypt_passwords(old_cipher, new_cipher):
+    for root, _, files in os.walk(DATA_DIR):
+        for file in files:
+            if file.endswith('.pass'):
+                file_path = Path(root) / file
+                with open(file_path, 'r') as f:
+                    lines = f.read().splitlines()
+                    if len(lines) < 4:
+                        click.echo(f"Invalid password file format for {file_path.stem}. Skipping.")
+                        continue
+                    description = lines[1]
+                    user_id = lines[2]
+                    encrypted_password = lines[3].encode()
                     try:
-                        cls._instance._initialize_cipher()
-                    except FileNotFoundError:
-                        cls._instance.cipher = None
-        return cls._instance
+                        decrypted_password = old_cipher.decrypt(encrypted_password)
+                    except InvalidToken:
+                        click.echo(f"Failed to decrypt {file_path.stem}. Skipping.")
+                        continue
+                    new_encrypted_password = new_cipher.encrypt(decrypted_password)
+                    password_entry = f"{lines[0]}\n{description}\n{user_id}\n{new_encrypted_password.decode()}"
+                    with open(file_path, 'w') as f:
+                        f.write(password_entry)
+    logging.info("Re-encryption with new master password completed.")
 
-    def _initialize_cipher(self):
-        self.cipher = self._create_cipher()
-        self._cipher_initialized = True
+def decrypt_master_password():
+    """Decrypt and return the master password using the secure key."""
+    secure_key = load_secure_key()
+    fernet = Fernet(secure_key)
+    with open(MASTER_PASSWORD_FILE, 'rb') as f:
+        encrypted_password = f.read()
+    password = fernet.decrypt(encrypted_password).decode()
+    return password
 
-    def _create_cipher(self):
-        password = decrypt_master_password()
-        key = derive_key(password)
-        return Fernet(key)
-
-    def get_cipher(self):
-        if not self._cipher_initialized:
-            raise ValueError("Cipher not initialized. Run 'vault set-master-password' to set it.")
-        return self.cipher
-
-    def refresh_cipher(self):
-        with self._lock:
-            old_cipher = self.cipher
-            self.cipher = self._create_cipher()
-            self._cipher_initialized = True
-            reencrypt_passwords(old_cipher, self.cipher)
-
-# Instantiate the cipher singleton
-cipher_singleton = CipherSingleton()
