+++ vault.log
@@ -259,3 +259,72 @@
 2024-06-22 18:22:13,942 - INFO - Re-encryption with new master password completed.
 2024-06-22 18:22:18,273 - INFO - Logging initialized
 2024-06-22 18:22:18,392 - DEBUG - Derived key: b'ZJ-GgYWHa-TUdTQMXSHVICbifDjytszSGQGMSZYv-ic='
+2024-06-22 19:13:05,264 - INFO - Logging initialized
+2024-06-22 19:14:01,949 - INFO - Logging initialized
+2024-06-22 19:14:04,126 - INFO - Master password stored in encrypted form on disk.
+2024-06-22 19:15:03,887 - INFO - Logging initialized
+2024-06-22 19:15:06,111 - INFO - Master password stored in encrypted form on disk.
+2024-06-22 19:15:06,159 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-22 19:15:06,159 - INFO - Re-encryption with new master password completed.
+2024-06-22 19:15:06,206 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-22 19:15:06,206 - INFO - Re-encryption with new master password completed.
+2024-06-22 19:15:11,263 - INFO - Logging initialized
+2024-06-22 19:15:11,385 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-22 19:15:16,268 - INFO - Logging initialized
+2024-06-22 19:15:16,388 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-22 19:15:18,321 - INFO - Logging initialized
+2024-06-22 19:15:18,441 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-22 19:15:18,490 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-22 19:15:18,491 - INFO - Re-encryption with new master password completed.
+2024-06-22 19:15:18,491 - INFO - Re-encryption with new master password completed.
+2024-06-22 19:15:20,748 - INFO - Logging initialized
+2024-06-22 19:15:20,867 - DEBUG - Derived key: b'lPZBQbZGZdUWEZv2phycrn9RBynQThNohzxdMJ8MWfQ='
+2024-06-23 20:16:15,401 - INFO - Logging initialized
+2024-06-23 20:16:17,332 - INFO - Master password stored in encrypted form on disk.
+2024-06-23 20:16:17,380 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:16:17,380 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:16:17,426 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:16:17,426 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:16:23,356 - INFO - Logging initialized
+2024-06-23 20:16:23,478 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:16:23,527 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:16:23,527 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:16:23,527 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:16:37,275 - INFO - Logging initialized
+2024-06-23 20:16:37,400 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:16:51,788 - INFO - Logging initialized
+2024-06-23 20:16:51,913 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:16:53,890 - INFO - Master password stored in encrypted form on disk.
+2024-06-23 20:16:53,951 - DEBUG - Derived key: b'nEVTMzuZNqUJLnpTBlhuOaijqE-nqr86_J77aMr4RlI='
+2024-06-23 20:16:53,951 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:16:53,998 - DEBUG - Derived key: b'nEVTMzuZNqUJLnpTBlhuOaijqE-nqr86_J77aMr4RlI='
+2024-06-23 20:16:53,998 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:16:55,804 - INFO - Logging initialized
+2024-06-23 20:16:55,921 - DEBUG - Derived key: b'nEVTMzuZNqUJLnpTBlhuOaijqE-nqr86_J77aMr4RlI='
+2024-06-23 20:17:02,612 - INFO - Logging initialized
+2024-06-23 20:17:02,732 - DEBUG - Derived key: b'nEVTMzuZNqUJLnpTBlhuOaijqE-nqr86_J77aMr4RlI='
+2024-06-23 20:17:04,841 - INFO - Master password stored in encrypted form on disk.
+2024-06-23 20:17:04,892 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:17:04,892 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:17:04,939 - DEBUG - Derived key: b'GUCxNcE4JetxL4bN1j_kasWXJAFbcifzvbzciUHXxMs='
+2024-06-23 20:17:04,939 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:20:34,275 - INFO - Logging initialized
+2024-06-23 20:31:52,285 - INFO - Logging initialized
+2024-06-23 20:32:19,826 - INFO - Logging initialized
+2024-06-23 20:32:21,777 - INFO - Master password stored in encrypted form on disk.
+2024-06-23 20:32:21,825 - DEBUG - Derived key: b'TZrnhmaQ7XpMf_psdyflDLyrGZbIxKpsja3AiYUJls8='
+2024-06-23 20:32:21,825 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:32:21,872 - DEBUG - Derived key: b'TZrnhmaQ7XpMf_psdyflDLyrGZbIxKpsja3AiYUJls8='
+2024-06-23 20:32:21,872 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:33:12,243 - INFO - Logging initialized
+2024-06-23 20:33:14,187 - INFO - Master password stored in encrypted form on disk.
+2024-06-23 20:33:14,239 - DEBUG - Derived key: b'3Nh3OtqZa11UqBpmh5qg4CcR92LjSuuiJTLW6kzX_dU='
+2024-06-23 20:33:14,239 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:33:14,289 - DEBUG - Derived key: b'3Nh3OtqZa11UqBpmh5qg4CcR92LjSuuiJTLW6kzX_dU='
+2024-06-23 20:33:14,289 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:33:42,467 - INFO - Logging initialized
+2024-06-23 20:33:44,875 - INFO - Master password stored in encrypted form on disk.
+2024-06-23 20:33:44,924 - DEBUG - Derived key: b'kKSMqVOgS8lgj7P6VxobK6wDDksrz8K6rhZGSKaUeR0='
+2024-06-23 20:33:44,924 - INFO - Re-encryption with new master password completed.
+2024-06-23 20:33:44,971 - DEBUG - Derived key: b'kKSMqVOgS8lgj7P6VxobK6wDDksrz8K6rhZGSKaUeR0='
+2024-06-23 20:33:44,971 - INFO - Re-encryption with new master password completed.
